♻️ refactor: AuthInput Component 파일 jsx > tsx 변경
